Don't get block light level if the sky level is 15
A very small light engine optimization. If the sky light level is 15, the block light level will never be higher,
since Minecraft doesn't support light levels above 15. This will prevent an extra call to getLightValue for
areas with sky access during the day. After some quick profiling on a flat world, this seems to reduce the time spent
on the getBrightness method by ~23%. The method is called every time a grass block randomly ticks
(along with crop growth), so this will cause a very slight reduction in the tick duration for most servers.